Rename and refactor modules for consistent naming.
