+++ data_management/story_image_clip.py
