+++ data_management/story_photo_backandwhite.py
