+++ data_management/story_photo_color_transformations.py
