+++ data_management/story_photo_grayscale.py
