+++ data_management/story_photo_transformer.py
@@ -170,11 +170,11 @@ class Application(PipelinePhase):
 
 if __name__ == "__main__":
     import story_image_clip
-    import story_photo_GUI_grayscale
-    import story_photo_GUI_backandwhite
+    import story_photo_grayscale
+    import story_photo_backandwhite
     import story_photo_removelines
 
-    phase_list = [Application, story_image_clip, story_photo_GUI_grayscale, story_photo_GUI_backandwhite,
+    phase_list = [Application, story_image_clip, story_photo_grayscale, story_photo_backandwhite,
                   story_photo_removelines]
 
     first = True
